Support: add local overloads for type-conversion
The imported values are imported as `Int32` rather than `UInt32` which
makes general usage of the values more difficult.  Provide a local
representation which uses the `UInt32` type.

+++ Sources/SwiftWin32/Views and Controls/DatePicker.swift
@@ -24,7 +24,7 @@ public class DatePicker: Control {
   private static let `class`: WindowClass =
       WindowClass(named: DATETIMEPICK_CLASS)
   private static let style: WindowStyle =
-      (base: WS_POPUP | DWORD(WS_TABSTOP), extended: 0)
+      (base: WS_POPUP | WS_TABSTOP, extended: 0)
 
   /// Configuring the Date Picker Style
